+++ aries_vcx/src/libindy/utils/anoncreds.rs
@@ -1002,7 +1002,7 @@ pub mod tests {
     use crate::libindy::utils::anoncreds::test_utils::{create_and_store_credential, create_and_store_credential_def, create_and_store_nonrevocable_credential, create_and_store_nonrevocable_credential_def, create_and_write_test_schema, create_proof, create_proof_with_predicate};
     use crate::utils::constants::SCHEMAS_JSON;
     use crate::utils::constants::TAILS_DIR;
-    use crate::utils::devsetup::{SetupLibraryWallet, SetupMocks, SetupWalletPoolAgency};
+    use crate::utils::devsetup::{SetupLibraryWallet, SetupMocks, SetupWalletPool};
     use crate::utils::get_temp_dir_path;
 
     use super::*;
@@ -1012,7 +1012,7 @@ pub mod tests {
     #[cfg(feature = "pool_tests")]
     #[tokio::test]
     async fn test_prover_verify_proof() {
-        let setup = SetupWalletPoolAgency::init().await;
+        let setup = SetupWalletPool::init().await;
 
         let (schemas, cred_defs, proof_req, proof) = create_proof(setup.wallet_handle).await;
 
@@ -1031,7 +1031,7 @@ pub mod tests {
     #[cfg(feature = "pool_tests")]
     #[tokio::test]
     async fn test_prover_verify_proof_with_predicate_success_case() {
-        let setup = SetupWalletPoolAgency::init().await;
+        let setup = SetupWalletPool::init().await;
 
         let (schemas, cred_defs, proof_req, proof) = create_proof_with_predicate(setup.wallet_handle, true).await;
 
@@ -1050,7 +1050,7 @@ pub mod tests {
     #[cfg(feature = "pool_tests")]
     #[tokio::test]
     async fn test_prover_verify_proof_with_predicate_fail_case() {
-        let setup = SetupWalletPoolAgency::init().await;
+        let setup = SetupWalletPool::init().await;
 
         let (schemas, cred_defs, proof_req, proof) = create_proof_with_predicate(setup.wallet_handle, false).await;
 
@@ -1096,7 +1096,7 @@ pub mod tests {
     #[cfg(feature = "pool_tests")]
     #[tokio::test]
     async fn test_issuer_revoke_credential() {
-        let setup = SetupWalletPoolAgency::init().await;
+        let setup = SetupWalletPool::init().await;
 
         let rc = libindy_issuer_revoke_credential(setup.wallet_handle, get_temp_dir_path(TAILS_DIR).to_str().unwrap(), "", "").await;
         assert!(rc.is_err());
@@ -1111,7 +1111,7 @@ pub mod tests {
     #[cfg(feature = "pool_tests")]
     #[tokio::test]
     async fn test_create_cred_def_real() {
-        let setup = SetupWalletPoolAgency::init().await;
+        let setup = SetupWalletPool::init().await;
 
         let (schema_id, _) = create_and_write_test_schema(setup.wallet_handle, utils::constants::DEFAULT_SCHEMA_ATTRS).await;
         let (_, schema_json) = get_schema_json(setup.wallet_handle, &schema_id).await.unwrap();
@@ -1125,7 +1125,7 @@ pub mod tests {
     #[tokio::test]
     async fn test_rev_reg_def_fails_for_cred_def_created_without_revocation() {
         // todo: does not need agency setup
-        let setup = SetupWalletPoolAgency::init().await;
+        let setup = SetupWalletPool::init().await;
 
         // Cred def is created with support_revocation=false,
         // revoc_reg_def will fail in libindy because cred_Def doesn't have revocation keys
@@ -1139,7 +1139,7 @@ pub mod tests {
     #[cfg(feature = "pool_tests")]
     #[tokio::test]
     async fn test_create_rev_reg_def() {
-        let setup = SetupWalletPoolAgency::init().await;
+        let setup = SetupWalletPool::init().await;
 
         let (schema_id, _) = create_and_write_test_schema(setup.wallet_handle, utils::constants::DEFAULT_SCHEMA_ATTRS).await;
         let (_, schema_json) = get_schema_json(setup.wallet_handle, &schema_id).await.unwrap();
@@ -1155,7 +1155,7 @@ pub mod tests {
     #[cfg(feature = "pool_tests")]
     #[tokio::test]
     async fn test_get_rev_reg_def_json() {
-        let setup = SetupWalletPoolAgency::init().await;
+        let setup = SetupWalletPool::init().await;
 
         let attrs = r#"["address1","address2","city","state","zip"]"#;
         let (_, _, _, _, rev_reg_id, _, _) = create_and_store_credential_def(setup.wallet_handle, attrs).await;
@@ -1167,7 +1167,7 @@ pub mod tests {
     #[cfg(feature = "pool_tests")]
     #[tokio::test]
     async fn test_get_rev_reg_delta_json() {
-        let setup = SetupWalletPoolAgency::init().await;
+        let setup = SetupWalletPool::init().await;
 
         let attrs = r#"["address1","address2","city","state","zip"]"#;
         let (_, _, _, _, rev_reg_id, _, _) = create_and_store_credential_def(setup.wallet_handle, attrs).await;
@@ -1179,7 +1179,7 @@ pub mod tests {
     #[cfg(feature = "pool_tests")]
     #[tokio::test]
     async fn test_get_rev_reg() {
-        let setup = SetupWalletPoolAgency::init().await;
+        let setup = SetupWalletPool::init().await;
 
         let attrs = r#"["address1","address2","city","state","zip"]"#;
         let (_, _, _, _, rev_reg_id, _, _) = create_and_store_credential_def(setup.wallet_handle, attrs).await;
@@ -1191,7 +1191,7 @@ pub mod tests {
     #[cfg(feature = "pool_tests")]
     #[tokio::test]
     async fn test_get_cred_def() {
-        let setup = SetupWalletPoolAgency::init().await;
+        let setup = SetupWalletPool::init().await;
 
         let attrs = r#"["address1","address2","city","state","zip"]"#;
         let (_, _, cred_def_id, cred_def_json, _) = create_and_store_nonrevocable_credential_def(setup.wallet_handle, attrs).await;
@@ -1204,7 +1204,7 @@ pub mod tests {
     #[cfg(feature = "pool_tests")]
     #[tokio::test]
     async fn test_is_cred_def_on_ledger() {
-        let _setup = SetupWalletPoolAgency::init().await;
+        let _setup = SetupWalletPool::init().await;
 
         assert_eq!(is_cred_def_on_ledger(None, "V4SGRU86Z58d6TV7PBUe6f:3:CL:194:tag7").await.unwrap(), false);
     }
@@ -1212,7 +1212,7 @@ pub mod tests {
     #[cfg(feature = "pool_tests")]
     #[tokio::test]
     async fn from_pool_ledger_with_id() {
-        let setup = SetupWalletPoolAgency::init().await;
+        let setup = SetupWalletPool::init().await;
 
         let (schema_id, _schema_json) = create_and_write_test_schema(setup.wallet_handle, utils::constants::DEFAULT_SCHEMA_ATTRS).await;
 
@@ -1235,7 +1235,7 @@ pub mod tests {
     #[cfg(feature = "pool_tests")]
     #[tokio::test]
     async fn test_revoke_credential() {
-        let setup = SetupWalletPoolAgency::init().await;
+        let setup = SetupWalletPool::init().await;
 
         let (_, _, _, _, _, _, _, _, rev_reg_id, cred_rev_id)
             = create_and_store_credential(setup.wallet_handle, utils::constants::DEFAULT_SCHEMA_ATTRS).await;
@@ -1257,7 +1257,7 @@ pub mod tests {
     #[cfg(feature = "pool_tests")]
     #[tokio::test]
     async fn test_get_txn() {
-        let setup = SetupWalletPoolAgency::init().await;
+        let setup = SetupWalletPool::init().await;
         get_ledger_txn(setup.wallet_handle, None, 0).await.unwrap_err();
         let txn = get_ledger_txn(setup.wallet_handle, None, 1).await;
         assert!(txn.is_ok());
